### 2018-01-02 #177 @arendsee
rhmmer - presubmission
Parses output of HMMER into tidy data frames

```
Package: rhmmer
Version: 0.1.0
Title: Utilities Parsing 'HMMER' Results
    'HMMER' is a profile hidden Markov model tool used primarily for sequence
    analysis in bioinformatics (<http://hmmer.org/>). 'rhmmer' provides
    utilities for parsing the 'HMMER' output into tidy data frames.
Authors@R: person("Zebulun", "Arendsee", email="zbwrnz@gmail.com", role=c("aut","cre"))
    rlang
    covr,
```

https://github.com/arendsee/rhmmer

data munging, because the package parses the idiosyncratic output of a tool into tidy data frames

Researchers in bioinformatics (especially genome annotation and sequence analysis) who want to analyze HMMER output in R.

yours differ or meet [our criteria for best-in-category]

There are no other packages for working with the standalone version of HMMER. There are packages for submitting jobs to the online HMMER server (e.g. `Bio3D`), though these are necessarily limited to low-throughput analyses.

It already is on CRAN. But I can move the repository if accepted.

	- [ ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.
