### 2018-01-02 #177 @arendsee
I submit this as a presubmission because I am not certain whether it is substantial enough for entry into ropensci. It is a very minimal tool specific for interfacing with one program.

However, I think there are several useful directions in which it could evolve, so I wanted to share it with the community.
