### 2018-01-17 #177 @karthik
Hi @arendsee
The editors and I feel like this would be a good fit for a full submission. However, we are keen to learn if you think there is similar class of file types that this package could be expanded to handle?
