### 2018-01-17 #177 @arendsee
Off the cuff, I can think two other formats.

 1. The pfam format that is the default output of `hmmscan`. This would be a pretty heavy parse job. I would probably want to implement it in C++/Rcpp.

 2. The HMM profiles. Here I would want to do both reading and writing. Then we could develop the profiles in R and write them for use in HMMER.

There are other formats produced by HMMER tools besides `hmmscan`. I will check on them and get back to you with more info.

HMMER is an incredibly awesome tool with cool applications even outside of bioinformatics (it has been used to compare bird songs, for example). What I have written is pretty minimal, but I hope to find other programmers who are interested in building on `rhmmer` to better expose the functionality of HMMER to the R world.

I can commit to maintaining the current functions of `rhmmer`, but I am unlikely to personally add any new features. Rather I'd try to find collaborators, ideally someone actually associated with the HMMER project.
