### 2018-01-17 #177 @arendsee
Here are a few other formats:

 1. `hmmstat` profile statistics output

 2. `hmmlogo` output used to build logo plots from a profile

 3. `hmmalign` output, this is just Stockholm format, I suspect parsers for this format already exist. But we could add a wrapper.

 All of the outputs of HMMER also contain metadata, which I am currently not parsing out.
